Fix warning about cast to pointer from integer of different size.
Make an intermediate cast to a intptr_t(an integer type that can hold a pointer)
before casting int or int *.

+++ C/list/test_deque.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdint.h>
 #include <stdlib.h>
 
 #include "deque.h"
@@ -16,10 +17,10 @@ int main(int argc, char **argv) {
   deque_push(&my_deque, (void *)3);
 
   int *a = (int *)deque_peek(&my_deque);
-  printf("Taking a peek at the front of the deque: %d\n", (int)a);
+  printf("Taking a peek at the front of the deque: %d\n", (int)(intptr_t)a);
 
   a = (int *)deque_peek_left(&my_deque);
-  printf("Taking a peek at the back of the deque: %d\n", (int)a);
+  printf("Taking a peek at the back of the deque: %d\n", (int)(intptr_t)a);
 
   printf("Push-ing left 1,2,3\n");
 
@@ -31,55 +32,55 @@ int main(int argc, char **argv) {
   printf("Pop-ing!\n"); 
 
   a = (int *)deque_peek(&my_deque);
-  printf("Taking a head at the front of the deque: %d\n", (int)a);
+  printf("Taking a head at the front of the deque: %d\n", (int)(intptr_t)a);
 
   a = (int *)deque_peek_left(&my_deque);
-  printf("Taking a peek at the back of the deque: %d\n", (int)a);
+  printf("Taking a peek at the back of the deque: %d\n", (int)(intptr_t)a);
 
   deque_pop_left(&my_deque);
   printf("Pop-ing left!\n"); 
 
   a = (int *)deque_peek(&my_deque);
-  printf("Taking a head at the front of the deque: %d\n", (int)a);
+  printf("Taking a head at the front of the deque: %d\n", (int)(intptr_t)a);
 
   a = (int *)deque_peek_left(&my_deque);
-  printf("Taking a peek at the back of the deque: %d\n", (int)a);
+  printf("Taking a peek at the back of the deque: %d\n", (int)(intptr_t)a);
 
   deque_pop(&my_deque);
   printf("Pop-ing!\n"); 
 
   a = (int *)deque_peek(&my_deque);
-  printf("Taking a head at the front of the deque: %d\n", (int)a);
+  printf("Taking a head at the front of the deque: %d\n", (int)(intptr_t)a);
 
   a = (int *)deque_peek_left(&my_deque);
-  printf("Taking a peek at the back of the deque: %d\n", (int)a);
+  printf("Taking a peek at the back of the deque: %d\n", (int)(intptr_t)a);
 
   deque_pop_left(&my_deque);
   printf("Pop-ing left!\n"); 
 
   a = (int *)deque_peek(&my_deque);
-  printf("Taking a head at the front of the deque: %d\n", (int)a);
+  printf("Taking a head at the front of the deque: %d\n", (int)(intptr_t)a);
 
   a = (int *)deque_peek_left(&my_deque);
-  printf("Taking a peek at the back of the deque: %d\n", (int)a);
+  printf("Taking a peek at the back of the deque: %d\n", (int)(intptr_t)a);
 
   deque_pop(&my_deque);
   printf("Pop-ing!\n"); 
 
   a = (int *)deque_peek(&my_deque);
-  printf("Taking a head at the front of the deque: %d\n", (int)a);
+  printf("Taking a head at the front of the deque: %d\n", (int)(intptr_t)a);
 
   a = (int *)deque_peek_left(&my_deque);
-  printf("Taking a peek at the back of the deque: %d\n", (int)a);
+  printf("Taking a peek at the back of the deque: %d\n", (int)(intptr_t)a);
 
   deque_pop_left(&my_deque);
   printf("Pop-ing left!\n"); 
 
   a = (int *)deque_peek(&my_deque);
-  printf("Taking a head at the front of the deque: %d\n", (int)a);
+  printf("Taking a head at the front of the deque: %d\n", (int)(intptr_t)a);
 
   a = (int *)deque_peek_left(&my_deque);
-  printf("Taking a peek at the back of the deque: %d\n", (int)a);
+  printf("Taking a peek at the back of the deque: %d\n", (int)(intptr_t)a);
 
   deque_pop(&my_deque);
